Push six new Kattis challenges, one additional to be completed

+++ open.kattis.com/Rocky Mountain Regional Contest (RMRC) 2016/stacking_cups.cpp
@@ -9,9 +9,9 @@ int main() {
 	int i = 0;
 	while(i < N) {
 		std::string C, tmpR;
-		int R;
-		
+		double R;
 		std::cin >> C >> tmpR;
+		
 		try {
 			R = std::stod(tmpR);
 		} catch(const std::exception&) {
